requirements.txt
streamlit 
transformers 
torch 
torchvision
Pillow 
